Restructured student views, added docstrings and constant
