+++ portal/forms/teach.py
@@ -359,7 +359,7 @@ class TeacherSetStudentPass(forms.Form):
         password = self.cleaned_data.get("password", None)
 
         if password and not password_strength_test(
-            password, length=8, upper=False, lower=False, numbers=False
+            password, upper=False, lower=False, numbers=False
         ):
             raise forms.ValidationError(
                 "Password not strong enough, consider using at least 8 characters, upper and lower case letters, and numbers"
